Moved callback error redirect settings parser/loader to validation package Fixed error style for lint pass
